Fix variable type.

+++ src/database/engine/rrdengineapi.c
@@ -59,7 +59,7 @@ __attribute__((constructor)) void initialize_multidb_ctx(void) {
         initialize_single_ctx(multidb_ctx[i]);
 }
 
-size_t dbengine_out_of_memory_protection = 0;
+uint64_t dbengine_out_of_memory_protection = 0;
 bool dbengine_use_all_ram_for_caches = false;
 int db_engine_journal_check = 0;
 bool new_dbengine_defaults = false;
